refactor: 용도에 따른 상수 분리
