+++ src/main/java/baseball/util/constant/NumberConstant.java
@@ -1,6 +1,9 @@
 package baseball.util.constant;
 
 public class NumberConstant {
-    public static final int MIN = 123;
-    public static final int MAX = 987;
+    public static final int INPUT_MIN = 123;
+    public static final int INPUT_MAX = 987;
+    public static final int RANDOM_NUMBER_SIZE = 3;
+    public static final int RANDOM_NUMBER_MIN = 1;
+    public static final int RANDOM_NUMBER_MAX = 9;
 }
